+++ x509-certificate/CHANGELOG.md
@@ -10,6 +10,10 @@ Released on ReleaseDate.
   `Zeroize<Vec<u8>>` instead of a `Vec<u8>`.
 * `InMemorySigningKeyPair` variants are now structs instead of
   unnamed tuples.
+* `InMemorySigningKeyPair` no longer implements
+  `From<ring::signature::Ed25519KeyPair>`.
+* `InMemorySigningKeyPair` now has a `to_pkcs8_one_asymmetric_key_der()` method
+  to obtain this PKCS#8 DER representation of the key pair.
 
 ## 0.22.1